Added Folder class and List of Old Addresses
Folder class to refactor some of the repeated path handling throughout. Added function to get a list of all of the addresses older than 10 days, potentially to send out weekly updates to coords or send directly to Bryan at the end of the week. Getting age of the report is fine, but the problemReport() method is returning 'UnboundLocalError: local variable 'Var1' referenced before assignment' for some of the files. Not strictly necessary, but it would be handy to keep that info in the object.